requirements.txt
matplotlib==3.5.1
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.5.0
segment_anything==1.0
